(QENG-1037) Install from git should accept depth
When installing from larger repositories, being able to specify depth can cut
the time of git checkout in half. This becomes especially helpful when you are
checking out multiple repositories.
Due to the older git on some of the templates, this provides the older
--branch name --depth 1 commands, which means when using depth, one should
ensure the rev passed is a branch and not a single commit. Alternatively one
can add depth_branch => 'name' to repository and have that used instead of rev.